Review Request 15326: Update test after schema update.

+++ le-pls/src/test/java/com/latticeengines/pls/end2end/CSVFileImportDeploymentTestNG.java
@@ -352,8 +352,8 @@ public class CSVFileImportDeploymentTestNG extends CDLDeploymentTestNGBase {
         Schema schema = AvroUtils.getSchema(yarnConfiguration, new Path(avroFiles.get(0)));
         Assert.assertEquals(schema.getField("TransactionTime").schema().getTypes().get(0).getType(),
                 Schema.Type.STRING);
-        Assert.assertEquals(schema.getField("Amount").schema().getTypes().get(0).getType(), Schema.Type.INT);
-        Assert.assertEquals(schema.getField("Quantity").schema().getTypes().get(0).getType(), Schema.Type.INT);
+        Assert.assertEquals(schema.getField("Amount").schema().getTypes().get(0).getType(), Schema.Type.DOUBLE);
+        Assert.assertEquals(schema.getField("Quantity").schema().getTypes().get(0).getType(), Schema.Type.DOUBLE);
     }
 
     @Test(groups = "deployment")
